# WinCCOAGraphQLRestApisExample/javascript/winccoa-graphql-server-main/restapi/openapi-full.yaml
openapi: 3.0.0
info:
  title: WinCC OA REST API
  version: 1.0.0
  description: |
    REST API for WinCC OA GraphQL Server - provides HTTP endpoints for all WinCC OA operations including:
    - Data point management (CRUD, value operations)
    - Data point type management
    - Tag queries with metadata
    - Alert handling
    - CNS (Central Navigation Service) operations
    - System information and redundancy
    - OPC UA configuration
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /restapi/v1/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing required parameter
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ElementType:
      type: string
      enum:
        - BOOL
        - UINT8
        - INT32
        - INT64
        - FLOAT
        - DOUBLE
        - BIT
        - BIT32
        - BIT64
        - STRING
        - TIME
        - DPID
        - LANGSTRING
        - BLOB
        - MIXED
        - DYN_BOOL
        - DYN_UINT8
        - DYN_INT32
        - DYN_INT64
        - DYN_FLOAT
        - DYN_DOUBLE
        - DYN_BIT
        - DYN_BIT32
        - DYN_BIT64
        - DYN_STRING
        - DYN_TIME
        - DYN_DPID
        - DYN_LANGSTRING
        - DYN_BLOB
    CtrlType:
      type: string
      enum:
        - TIME_VAR
        - BOOL_VAR
        - INT_VAR
        - UINT_VAR
        - FLOAT_VAR
        - STRING_VAR
        - BIT32_VAR
        - CHAR_VAR
        - DYN_TIME_VAR
        - DYN_BOOL_VAR
        - DYN_INT_VAR
        - DYN_UINT_VAR
        - DYN_FLOAT_VAR
        - DYN_STRING_VAR
        - DYN_BIT32_VAR
        - DYN_CHAR_VAR
        - ATIME_VAR
        - LANGSTRING_VAR
        - BLOB_VAR
        - LONG_VAR
        - ULONG_VAR
        - BIT64_VAR
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
    Forbidden:
      description: Insufficient permissions (admin role required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Admin role required
paths:
...
  /restapi/v1/datapoints:
    get:
      tags: [Data Points]
      summary: List data points
      description: Search for data points by pattern
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
          description: Search pattern (wildcards supported)
          example: "ExampleDP_*"
        - name: dpType
          in: query
          schema:
            type: string
          description: Filter by data point type
        - name: ignoreCase
          in: query
          schema:
            type: boolean
          description: Case-insensitive search
      responses:
        '200':
          description: List of data points
          content:
            application/json:
              schema:
                type: object
                properties:
                  datapoints:
                    type: array
                    items:
                      type: string
                    example: ["ExampleDP_Arg1.", "ExampleDP_Arg2."]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Data Points]
      summary: Create data point
      description: Create a new data point (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dpeName, dpType]
              properties:
                dpeName:
                  type: string
                  example: "MyDataPoint."
                dpType:
                  type: string
                  example: "ExampleDP_Arg"
                systemId:
                  type: integer
                  example: 1
                dpId:
                  type: integer
                  example: 100
      responses:
        '200':
          description: Data point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Data Points - Value Operations
  /restapi/v1/datapoints/{dpeName}/value:
    get:
      tags: [Data Points]
      summary: Get data point value
      description: Retrieve current value of a data point element
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
          description: Data point element name (URL encoded)
          example: "ExampleDP_Arg1.value"
      responses:
        '200':
          description: Current value
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    description: Current value (any type)
                    example: 42.5

    put:
      tags: [Data Points]
      summary: Set data point value
      description: Set the value of a data point element (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
          description: Data point element name (URL encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  description: New value (any type)
                  example: 50.0
      responses:
        '200':
          description: Value set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /restapi/v1/datapoints/{dpeName}/value/wait:
    put:
      tags: [Data Points]
      summary: Set value with wait
      description: Set value and wait for confirmation (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  example: 50.0
      responses:
        '200':
          description: Value set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoints/{dpeName}/value/timed:
    put:
      tags: [Data Points]
      summary: Set value with timestamp
      description: Set value with a specific timestamp (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, time]
              properties:
                value:
                  example: 50.0
                time:
                  type: string
                  format: date-time
                  example: "2024-01-01T12:00:00Z"
      responses:
        '200':
          description: Value set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoints/{dpeName}/value/timed-wait:
    put:
      tags: [Data Points]
      summary: Set value with timestamp and wait
      description: Set value with timestamp and wait for confirmation (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, time]
              properties:
                value:
                  example: 50.0
                time:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Value set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoints/{dpName}:
    delete:
      tags: [Data Points]
      summary: Delete data point
      description: Delete a data point (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Data point deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoints/{source}/copy:
    post:
      tags: [Data Points]
      summary: Copy data point
      description: Copy a data point (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: source
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination]
              properties:
                destination:
                  type: string
                  example: "NewDataPoint."
                driver:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Data point copied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoints/{dpeName}/exists:
    get:
      tags: [Data Points]
      summary: Check if data point exists
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Existence check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
  /restapi/v1/datapoints/{dpeName}/type:
    get:
      tags: [Data Points]
      summary: Get element type
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Element type
          content:
            application/json:
              schema:
                type: object
                properties:
                  elementType:
                    $ref: '#/components/schemas/ElementType'
  /restapi/v1/datapoints/{dpeName}/dp-type:
    get:
      tags: [Data Points]
      summary: Get data point type name
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Data point type
          content:
            application/json:
              schema:
                type: object
                properties:
                  dpType:
                    type: string
                    example: "ExampleDP_Arg"
  /restapi/v1/datapoints/{dpeName}/type-ref:
    get:
      tags: [Data Points]
      summary: Get type reference name
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Type reference
          content:
            application/json:
              schema:
                type: object
                properties:
                  typeRef:
                    type: string
  /restapi/v1/datapoints/{dpeName}/value/max-age:
    get:
      tags: [Data Points]
      summary: Get value with max age
      description: Get value from driver if older than specified age
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
        - name: age
          in: query
          required: true
          schema:
            type: integer
          description: Maximum age in milliseconds
          example: 1000
      responses:
        '200':
          description: Value retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    example: 42.5
  /restapi/v1/datapoints/{dpeName}/history:
    get:
      tags: [Data Points]
      summary: Get historic values
      description: Get historic values for a time period
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
        - name: startTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: endTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
          example: "2024-01-01T23:59:59Z"
      responses:
        '200':
          description: Historic values
          content:
            application/json:
              schema:
                type: object
                properties:
                  values:
                    type: array
                    items:
                      type: object
  /restapi/v1/datapoints/{dpAttributeName}/attribute-type:
    get:
      tags: [Data Points]
      summary: Get attribute type
      security:
        - bearerAuth: []
      parameters:
        - name: dpAttributeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attribute type
          content:
            application/json:
              schema:
                type: object
                properties:
                  ctrlType:
                    $ref: '#/components/schemas/CtrlType'
...